+++ starsky/starsky.foundation.platform/Extensions/ContentSecurityPolicyExtensions.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Builder;
+using starskycore.Middleware;
+
+namespace starsky.foundation.platform.Extensions
+{
+    public static class ContentSecurityPolicyExtensions
+    {
+        public static IApplicationBuilder UseContentSecurityPolicy(this IApplicationBuilder builder)
+        {
+            return builder.UseMiddleware<ContentSecurityPolicyMiddleware>();
+        }
+    }
+}
