+++ starsky/starsky.foundation.platform/Middleware/ContentSecurityPolicyMiddleware.cs
