+++ starsky/starsky.foundation.platform/starsky.foundation.platform.csproj
@@ -11,6 +11,7 @@
 
     <ItemGroup>
       <PackageReference Include="Microsoft.ApplicationInsights" Version="2.14.0" />
+      <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
       <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
       <PackageReference Include="Microsoft.Extensions.Configuration" Version="3.1.8" />
       <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="3.1.8" />
