+++ starsky/starskycore/Helpers/ConfigCliAppsStartupHelper.cs
@@ -8,7 +8,7 @@ using starsky.feature.import.Services;
 using starsky.foundation.database.Data;
 using starsky.foundation.database.Query;
 using starsky.foundation.injection;
-using starsky.foundation.platform.Middleware;
+using starsky.foundation.platform.Extensions;
 using starsky.foundation.platform.Models;
 using starsky.foundation.readmeta.Services;
 using starsky.foundation.storage.Interfaces;
