+++ starsky/starskycore/starskycore.csproj
@@ -25,13 +25,6 @@
     </ItemGroup>
 
     <ItemGroup>
-<!--        <PackageReference Include="Microsoft.Extensions.Configuration" Version="3.1.5" />-->
-<!--        <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="3.1.5" />-->
-<!--        <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="3.1.5" />-->
-<!--        <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="3.1.5" />-->
-    </ItemGroup>
-
-    <ItemGroup>
 <!--        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.5" />-->
 <!--        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="3.1.5" />-->
 <!--        <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="3.1.5" />-->
