+++ starsky/starskytest/Controllers/GeoControllerTest.cs
@@ -13,7 +13,7 @@ using starsky.feature.geolookup.Services;
 using starsky.foundation.database.Data;
 using starsky.foundation.database.Interfaces;
 using starsky.foundation.database.Query;
-using starsky.foundation.platform.Middleware;
+using starsky.foundation.platform.Extensions;
 using starskycore.Services;
 using starsky.foundation.platform.Models;
 using starsky.foundation.readmeta.Interfaces;
