+++ starsky/starskytest/Controllers/MetaUpdateControllerTest.cs
@@ -16,8 +16,8 @@ using starsky.foundation.database.Data;
 using starsky.foundation.database.Interfaces;
 using starsky.foundation.database.Models;
 using starsky.foundation.database.Query;
+using starsky.foundation.platform.Extensions;
 using starsky.foundation.platform.Helpers;
-using starsky.foundation.platform.Middleware;
 using starsky.foundation.platform.Models;
 using starsky.foundation.readmeta.Interfaces;
 using starsky.foundation.storage.Interfaces;
