+++ starsky/starskytest/Middleware/ContentSecurityPolicyMiddlewareTest.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using starsky.foundation.accountmanagement.Middleware;
+using starsky.foundation.platform.Extensions;
 using starskycore.Middleware;
 
 namespace starskytest.Middleware
