+++ starsky/starskytest/Services/BackgroundTaskQueueTest.cs
@@ -8,8 +8,8 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using starsky.foundation.platform.Extensions;
 using starsky.foundation.platform.Interfaces;
-using starsky.foundation.platform.Middleware;
 using starsky.foundation.platform.Models;
 using starskycore.Services;
 using starskytest.FakeMocks;
