+++ starsky/starskytest/starsky.foundation.platform/Helpers/ArgsHelperTest.cs
@@ -5,8 +5,8 @@ using System.Linq;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using starsky.foundation.platform.Extensions;
 using starsky.foundation.platform.Helpers;
-using starsky.foundation.platform.Middleware;
 using starsky.foundation.platform.Models;
 using starskycore.Attributes;
 using starskytest.FakeCreateAn;
